Organized the classes into suitable java packages.
