+++ src/main/java/Naega/Command/Command.java
@@ -1,3 +1,10 @@
+package Naega.Command;
+
+import Naega.NaegaException;
+import Naega.Storage.Storage;
+import Naega.Task.TaskList;
+import Naega.Ui.Ui;
+
 import java.io.IOException;
 
 public abstract class Command {
