+++ src/main/java/Naega/Command/DeleteCommand.java
@@ -1,3 +1,12 @@
+package Naega.Command;
+
+import Naega.Command.Command;
+import Naega.NaegaException;
+import Naega.Storage.Storage;
+import Naega.Task.Task;
+import Naega.Task.TaskList;
+import Naega.Ui.Ui;
+
 public class DeleteCommand extends Command {
     private final int taskIndex;
