+++ src/main/java/Naega/NaegaException.java
@@ -1,3 +1,5 @@
+package Naega;
+
 public class NaegaException extends Exception {
     public NaegaException(String message) {
         super(message);
