+++ src/main/java/Naega/Task/Task.java
@@ -1,3 +1,5 @@
+package Naega.Task;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
