+++ src/main/java/Naega/Task/Todo.java
@@ -1,3 +1,7 @@
+package Naega.Task;
+
+import Naega.Task.Task;
+
 public class Todo extends Task {
 
     public Todo(String description) {
