+++ src/main/java/Naega/Task/tasks.txt
@@ -1,0 +1,2 @@
+T | 0 | work
+E | 0 | be happy | 11/9/2024 2100 | 11/9/2090 2100
