+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
- Hello! I'm Naega
+ Hello! I'm Naega.Naega
  What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
